refactor: rename folder resources_NestJS
