+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersModule } from './resources_nestJS/users/users.module';
+import { UsersModule } from './resources/users/users.module';
 import ConnectionOptions from './common/ConnectionOptions';
-import { BoardsModule } from './resources_nestJS/boards/boards.module';
-import { TasksModule } from './resources_nestJS/tasks/tasks.module';
-import { LoginModule } from './resources_nestJS/login/login.module';
+import { BoardsModule } from './resources/boards/boards.module';
+import { TasksModule } from './resources/tasks/tasks.module';
+import { LoginModule } from './resources/login/login.module';
 
 @Module({
   imports: [
