+++ src/resources/users/interfaces/IReturnUser.ts
